refactor: cat-file 명령어 관련 util 함수들 분리
- CatFileUtil.kt에 분리
